restructure code as it was a pain
