+++ source/vector.cpp
