+++ source/vector.hpp
@@ -1,8 +1,5 @@
-#include <exception>
-
 #ifndef VECTOR_HPP
 #define VECTOR_HPP
-
 
 namespace cpp_utils {
